# swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Data Modeling Service",
        "description": "Data Modeling Service",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:3011",
            "description": "Local server url"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "paths": {
        "/api/tenants/{code}": {
            "get": {
                "tags": [
                    "Tenant"
                ],
                "operationId": "fetchByCode",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseTenantRes"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponseMapStringString"
                                        },
                                        {
                                            "$ref": "#/components/schemas/ApiResponseErrorDetail"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorDetail"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorDetail"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorDetail"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Tenant"
                ],
                "operationId": "update",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TenantDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseTenantRes"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponseMapStringString"
                                        },
                                        {
                                            "$ref": "#/components/schemas/ApiResponseErrorDetail"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorDetail"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorDetail"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorDetail"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tenant"
                ],
                "operationId": "deleteByCode",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseTenantRes"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponseMapStringString"
                                        },
                                        {
                                            "$ref": "#/components/schemas/ApiResponseErrorDetail"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorDetail"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorDetail"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorDetail"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/culture/blocks/{code}": {
            "get": {
                "tags": [
                    "CultureBlock"
                ],
                "summary": "Block Details",
                "operationId": "fetchByCode_1",
                "parameters": [
                    {
                        "name": "tenant",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "code",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseCultureBlock"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponseMapStringString"
                                        },
                                        {
                                            "$ref": "#/components/schemas/ApiResponseErrorDetail"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorDetail"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorDetail"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorDetail"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "CultureBlock"
                ],
                "summary": "Update Block",
                "operationId": "updateByCode",
                "parameters": [
                    {
                        "name": "tenant",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "code",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CultureBlockDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseVoid"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponseMapStringString"
                                        },
                                        {
                                            "$ref": "#/components/schemas/ApiResponseErrorDetail"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorDetail"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorDetail"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorDetail"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "CultureBlock"
                ],
                "summary": "Delete Block",
                "operationId": "deleteByCode_1",
                "parameters": [
                    {
                        "name": "tenant",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "code",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseVoid"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponseMapStringString"
                                        },
                                        {
                                            "$ref": "#/components/schemas/ApiResponseErrorDetail"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorDetail"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorDetail"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorDetail"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApiResponseErrorDetail": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "result": {
                        "$ref": "#/components/schemas/ErrorDetail"
                    }
                }
            },
            "ErrorDetail": {
                "type": "object",
                "properties": {
                    "errorCode": {
                        "type": "string"
                    },
                    "errorMessage": {
                        "type": "string"
                    }
                }
            },
            "ApiResponseMapStringString": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "result": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "string"
                        }
                    }
                }
            },
            "CohortItem": {
                "type": "object",
                "properties": {
                    "position": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "field": {
                        "type": "string"
                    },
                    "processor": {
                        "type": "string",
                        "enum": [
                            "YEAR_RANGE_FROM_YEAR",
                            "YEAR_RANGE_FROM_MONTHS",
                            "YEAR_RANGE_FROM_DATE",
                            "NONE"
                        ]
                    },
                    "active": {
                        "type": "boolean"
                    },
                    "display_name": {
                        "type": "string"
                    }
                }
            },
            "Encryption": {
                "type": "object",
                "properties": {
                    "enabled": {
                        "type": "boolean"
                    }
                }
            },
            "MicroCulture": {
                "type": "object",
                "properties": {
                    "threshold": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "frequency": {
                        "type": "string",
                        "enum": [
                            "DAILY",
                            "WEEKLY",
                            "MONTHLY",
                            "QUARTERLY",
                            "HALF_YEARLY",
                            "YEARLY"
                        ]
                    }
                }
            },
            "TenantDTO": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "code": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "active": {
                        "type": "boolean"
                    },
                    "subscription": {
                        "type": "string"
                    },
                    "currency": {
                        "type": "string"
                    },
                    "cohorts": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CohortItem"
                        }
                    },
                    "employees": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "initialize": {
                        "type": "boolean"
                    },
                    "median_salary": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "cost_of_disengagement": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "micro_culture_config": {
                        "$ref": "#/components/schemas/MicroCulture"
                    },
                    "encryption_config": {
                        "$ref": "#/components/schemas/Encryption"
                    }
                }
            },
            "ApiResponseTenantRes": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "result": {
                        "$ref": "#/components/schemas/TenantRes"
                    }
                }
            },
            "TenantRes": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "code": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "active": {
                        "type": "boolean"
                    },
                    "subscription": {
                        "type": "string"
                    },
                    "currency": {
                        "type": "string"
                    },
                    "cohorts": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CohortItem"
                        }
                    },
                    "employees": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "median_salary": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "cost_of_disengagement": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "micro_culture_config": {
                        "$ref": "#/components/schemas/MicroCulture"
                    },
                    "encryption_config": {
                        "$ref": "#/components/schemas/Encryption"
                    }
                }
            },
...
            "ApiResponseVoid": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "result": {
                        "type": "object"
                    }
                }
            },
            "CultureBlockDTO": {
                "type": "object",
                "properties": {
                    "tenant": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "code": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "ENPS",
                            "ENGAGEMENT",
                            "EXPERIENCE",
                            "PULSE",
                            "RUBICA_CONVERSATION",
                            "DEFAULT"
                        ]
                    },
                    "description": {
                        "type": "string"
                    },
                    "target": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "engines": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/WeightedCultureEngineDTO"
                        }
                    },
                    "active": {
                        "type": "boolean"
                    },
                    "questions": {
                        "uniqueItems": true,
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Question"
                        }
                    },
                    "formula": {
                        "type": "string"
                    }
                }
            },
            "Question": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "poolQuesReferenceCode": {
                        "type": "string"
                    },
                    "position": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "skippable": {
                        "type": "boolean"
                    },
                    "weight": {
                        "type": "number",
                        "format": "double"
                    },
                    "max": {
                        "type": "number",
                        "format": "double"
                    }
                }
            },
            "WeightedCultureEngineDTO": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "code": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "weight": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
...
            "ApiResponseCultureBlock": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "result": {
                        "$ref": "#/components/schemas/CultureBlock"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}